+++ sudokuru/app/Components/Sudoku Board/EndGameModal.tsx
@@ -5,11 +5,11 @@ import {useNavigation} from "@react-navigation/native";
 
 const EndGameModal = (props: any) => {
 
-    const theme = useTheme();
-    const navigation: any = useNavigation();
-
     const size = useWindowDimensions();
     const reSize = Math.min(size.width, size.height);
+
+    const theme = useTheme();
+    const navigation: any = useNavigation();
 
     const formatTime = (seconds: number) => {
         // Get minutes and remaining seconds
@@ -28,7 +28,7 @@ const EndGameModal = (props: any) => {
                 <View style={{ backgroundColor: '#fff', borderRadius: 10, padding: 20 }}>
                     <View style={{ marginBottom: 10 }}>
                         <Text style={{ fontSize: reSize/22, color: '#025E73'}}>Score:</Text>
-                        <Text style={{ fontSize: reSize/20, fontWeight: 'bold', color: '#D9A05B' }}>{formatTime(props.score)}</Text>
+                        <Text style={{ fontSize: reSize/20, fontWeight: 'bold', color: '#D9A05B' }}>{props.score}</Text>
                     </View>
                     <View style={{ marginBottom: 10 }}>
                         <Text style={{ fontSize: reSize/22, color: '#025E73'}}>Time Spent Playing:</Text>
@@ -47,7 +47,7 @@ const EndGameModal = (props: any) => {
                         <Text style={{ fontSize: reSize/20, fontWeight: 'bold', color: '#D9A05B' }}>{props.difficulty}</Text>
                     </View>
                 </View>
-            <Button mode="contained" onPress={navigation.navigate("Home")}>
+            <Button mode="contained" onPress={() => navigation.navigate("Home")}>
                 Go Home
             </Button>
         </View>
